requirements.txt
aiohttp>=3.8.0,<4.0.0
quart>=0.19.0,<1.0.0
quart-cors>=0.6.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
